Feat: align AMP to BF16-only and audit/fix all docs
Training changes:
- Remove --amp-dtype CLI flag, AMP_FP16_EPS constant, and all FP16/GradScaler
  branching; autocast hardcoded to bfloat16 everywhere (matches MIST)

Doc fixes:
- advanced_topics: features_path now correctly cites misfit_encode (not misfit_embed)
- advanced_topics: immutable parameter key corrected to model.architecture
- advanced_topics: --overwrite description no longer claims files are deleted upfront
- advanced_topics: spacing bullet rewritten to reflect native SpacingEmbedding conditioning
- advanced_topics/getting_started/README: GPU requirement now explicitly calls out Ampere+

Tests: 397 pass, 100% coverage

+++ misfit/evaluation/evaluator.py
@@ -170,7 +170,7 @@ class ReconstructionEvaluator:
         tensor = torch.from_numpy(patch).unsqueeze(0).unsqueeze(0)  # (1,1,D,H,W)
         tensor = tensor.to(self.device)
 
-        amp_ctx = torch.amp.autocast("cuda") if self.amp else nullcontext()
+        amp_ctx = torch.amp.autocast("cuda", dtype=torch.bfloat16) if self.amp else nullcontext()
         with torch.no_grad(), amp_ctx:
             output = self.model(tensor)
